+++ plugin-build/src/test/kotlin/io/sentry/android/gradle/tasks/SentryGenerateProguardUuidTaskTest.kt
@@ -1,7 +1,6 @@
 package io.sentry.android.gradle.tasks
 
 import java.io.File
-import kotlin.test.assertEquals
 import kotlin.test.assertNotEquals
 import kotlin.test.assertTrue
 import org.gradle.api.Project
@@ -12,23 +11,6 @@ import org.junit.Test
 class SentryGenerateProguardUuidTaskTest {
 
     @Test
-    fun `generate proguard UUID sets the output file correctly`() {
-        val project = createProject()
-        val task: TaskProvider<SentryGenerateProguardUuidTask> =
-            project.tasks.register(
-                "testGenerateProguardUuid",
-                SentryGenerateProguardUuidTask::class.java
-            ) {
-                it.outputDirectory.set(project.file("dummy/folder/"))
-            }
-
-        assertEquals(
-            project.file("dummy/folder/sentry-debug-meta.properties"),
-            task.get().outputFile.get().asFile
-        )
-    }
-
-    @Test
     fun `generate proguard UUID generates the UUID correctly`() {
         val project = createProject()
         val task: TaskProvider<SentryGenerateProguardUuidTask> =
@@ -36,7 +18,7 @@ class SentryGenerateProguardUuidTaskTest {
                 "testGenerateProguardUuid",
                 SentryGenerateProguardUuidTask::class.java
             ) {
-                it.outputDirectory.set(project.file("dummy/folder/"))
+                it.output.set(project.file("dummy/folder/sentry-debug-meta.properties"))
             }
 
         task.get().generateProperties()
@@ -54,7 +36,7 @@ class SentryGenerateProguardUuidTaskTest {
                 "testGenerateProguardUuid",
                 SentryGenerateProguardUuidTask::class.java
             ) {
-                it.outputDirectory.set(project.file("dummy/folder/"))
+                it.output.set(project.file("dummy/folder/sentry-debug-meta.properties"))
             }
         val expectedFile = File(project.projectDir, "dummy/folder/sentry-debug-meta.properties")